# frontend/openapi-s3-gateway.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "S3 Gateway",
        "version": "0.1.0"
    },
    "paths": {
...
        "/buckets/": {
            "post": {
                "summary": "Create Bucket",
                "operationId": "create_bucket_buckets__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BucketCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BucketResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/buckets/{bucket_id}/objects/": {
            "get": {
                "summary": "List Bucket Objects",
                "operationId": "list_bucket_objects_buckets__bucket_id__objects__get",
                "parameters": [
                    {
                        "name": "bucket_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Bucket Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BucketObjectListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/buckets/{bucket_id}/billing/": {
            "get": {
                "summary": "Get Bucket Billing",
                "operationId": "get_bucket_billing_buckets__bucket_id__billing__get",
                "parameters": [
                    {
                        "name": "bucket_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Bucket Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BillingResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BillingResponse": {
                "properties": {
                    "bucket_id": {
                        "type": "integer",
                        "title": "Bucket Id",
                        "description": "ID bucketu"
                    },
                    "bucket_name": {
                        "type": "string",
                        "title": "Bucket Name",
                        "description": "N\u00e1zev bucketu"
                    },
                    "bandwidth_bytes": {
                        "type": "integer",
                        "title": "Bandwidth Bytes",
                        "description": "Celkov\u00e9 p\u0159enesen\u00e9 bajty (legacy)"
                    },
                    "current_storage_bytes": {
                        "type": "integer",
                        "title": "Current Storage Bytes",
                        "description": "Aktu\u00e1ln\u00ed \u00falo\u017ei\u0161t\u011b v bajtech"
                    },
                    "ingress_bytes": {
                        "type": "integer",
                        "title": "Ingress Bytes",
                        "description": "P\u0159\u00edchoz\u00ed p\u0159enosy v bajtech"
                    },
                    "egress_bytes": {
                        "type": "integer",
                        "title": "Egress Bytes",
                        "description": "Odchoz\u00ed p\u0159enosy v bajtech"
                    },
                    "internal_transfer_bytes": {
                        "type": "integer",
                        "title": "Internal Transfer Bytes",
                        "description": "Intern\u00ed p\u0159enosy v bajtech"
                    }
                },
                "type": "object",
                "required": [
                    "bucket_id",
                    "bucket_name",
                    "bandwidth_bytes",
                    "current_storage_bytes",
                    "ingress_bytes",
                    "egress_bytes",
                    "internal_transfer_bytes"
                ],
                "title": "BillingResponse"
            },
...
            "BucketCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name",
                        "description": "N\u00e1zev bucketu"
                    }
                },
                "type": "object",
                "required": [
                    "name"
                ],
                "title": "BucketCreate"
            },
            "BucketObjectListResponse": {
                "properties": {
                    "bucket_id": {
                        "type": "integer",
                        "title": "Bucket Id",
                        "description": "ID bucketu"
                    },
                    "files": {
                        "items": {
                            "$ref": "#/components/schemas/FileMetadata"
                        },
                        "type": "array",
                        "title": "Files",
                        "description": "Seznam objekt\u016f v bucketu"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total",
                        "description": "Celkov\u00fd po\u010det objekt\u016f"
                    }
                },
                "type": "object",
                "required": [
                    "bucket_id",
                    "files",
                    "total"
                ],
                "title": "BucketObjectListResponse"
            },
            "BucketResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id",
                        "description": "ID bucketu"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name",
                        "description": "N\u00e1zev bucketu"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At",
                        "description": "\u010cas vytvo\u0159en\u00ed bucketu"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "created_at"
                ],
                "title": "BucketResponse"
            },
...
            "FileMetadata": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id",
                        "description": "Unik\u00e1tn\u00ed ID souboru"
                    },
                    "filename": {
                        "type": "string",
                        "title": "Filename",
                        "description": "P\u016fvodn\u00ed n\u00e1zev souboru"
                    },
                    "size": {
                        "type": "integer",
                        "title": "Size",
                        "description": "Velikost souboru v bytech"
                    },
                    "content_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Content Type",
                        "description": "MIME typ souboru"
                    },
                    "created_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "string"
                            }
                        ],
                        "title": "Created At",
                        "description": "\u010cas vytvo\u0159en\u00ed souboru"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "filename",
                    "size",
                    "created_at"
                ],
                "title": "FileMetadata"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}